requirements.txt
openai>=1.0.0
datasets
huggingface_hub
PyYAML
numpy
tqdm